+++ src/creatio/contracts/index.ts
@@ -2,6 +2,7 @@ export * from './admin-operation-provider';
 export * from './configuration-provider';
 export * from './crud-provider';
 export * from './feature-provider';
+export * from './file-provider';
 export * from './process-provider';
 export * from './query';
 export * from './sys-settings-provider';